refactor xxhash
