+++ Cargo.toml
@@ -201,7 +201,7 @@ url = { version = "2.5.4" }
 uuid = { version = "1.18.1", features = ["v4", "v8"] }
 walkdir = "2.5.0"
 which = { version = "8.0.0", features = [] }
-xxhash-rust = { version = "0.8.15", features = ["const_xxh3", "const_xxh32", "const_xxh64", "xxh3", "xxh32", "xxh64"] }
+xxhash-rust = { version = "0.8.15", features = []}
 zstd = "0.13.3"
 zstd-safe = "7.2.4"
